requirements.txt
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.19.2
python-nmap==0.7.1
rich==14.0.0
pyyaml==6.0.2
ipaddress==1.0.23
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
concurrent.futures==3.1.1
threading==0.3.8
smtplib==1.0.0
email-validator==2.1.0
jinja2==3.1.3
plotly==5.18.0
